Unit tests

+++ User-Service/src/main/java/org/n11/controller/contract/impl/UserControllerContractImpl.java
@@ -2,7 +2,7 @@ package org.n11.controller.contract.impl;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.aspectj.bridge.Message;
+import org.n11.constant.ErrorMessages;
 import org.n11.constant.Messages;
 import org.n11.controller.contract.UserControllerContract;
 import org.n11.entity.User;
@@ -12,10 +12,12 @@ import org.n11.entity.request.UserSaveRequest;
 import org.n11.entity.request.UserUpdateRequest;
 import org.n11.service.UserEntityService;
 import org.n11.service.mapper.UserMapper;
+import org.n11.utilities.exceptions.ItemNotFoundException;
 import org.n11.utilities.helper.BusinessRules.RegularExpression;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -34,7 +36,7 @@ public class UserControllerContractImpl implements UserControllerContract {
     public UserDTO save(UserSaveRequest userSaveRequest) {
         RegularExpression.controlEmail(userSaveRequest.email());
         RegularExpression.controlPhoneNumber(userSaveRequest.phoneNumber());
-        RegularExpression.controlFirstNameAndLastName(userSaveRequest.firstName(),userSaveRequest.lastName());
+        RegularExpression.controlFirstNameAndLastName(userSaveRequest.firstName(), userSaveRequest.lastName());
 
         User user = UserMapper.INSTANCE.convertToUser(userSaveRequest);
         this.userEntityService.save(user);
@@ -43,13 +45,16 @@ public class UserControllerContractImpl implements UserControllerContract {
 
     @Transactional
     @Override
-    public UserDTO update(Long id,UserUpdateRequest userUpdateRequest) {
-        User user=this.userEntityService.findByIdWithControl(userUpdateRequest.id());
+    public UserDTO update(UserUpdateRequest userUpdateRequest) {
+        User user = this.userEntityService.findByIdWithControl(userUpdateRequest.id());
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
         RegularExpression.controlEmail(userUpdateRequest.email());
         RegularExpression.controlPhoneNumber(userUpdateRequest.phoneNumber());
-        RegularExpression.controlFirstNameAndLastName(userUpdateRequest.firstName(),userUpdateRequest.lastName());
+        RegularExpression.controlFirstNameAndLastName(userUpdateRequest.firstName(), userUpdateRequest.lastName());
 
-        UserMapper.INSTANCE.updateUser(user,userUpdateRequest);
+        UserMapper.INSTANCE.updateUser(user, userUpdateRequest);
         this.userEntityService.save(user);
         return UserMapper.INSTANCE.convertToDTO(user);
 
@@ -57,9 +62,11 @@ public class UserControllerContractImpl implements UserControllerContract {
     }
 
     @Override
-    public UserDTO getUserById(Long id) {
-        User user= this.userEntityService.findByIdWithControl(id);
-
+    public UserDTO findById(Long id) {
+        User user = this.userEntityService.findByIdWithControl(id);
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
         return UserMapper.INSTANCE.convertToDTO(user);
     }
 
@@ -68,28 +75,62 @@ public class UserControllerContractImpl implements UserControllerContract {
         List<User> userList = this.userEntityService.findAll()
                 .stream()
                 .filter(user -> user.getStatus() == Status.ACTIVE)
-                .collect(Collectors.toList());
+                .toList();
 
         return UserMapper.INSTANCE.convertToDTOs(userList);
     }
 
     @Override
+    public List<UserDTO> findAllByDeactive() {
+        List<User> userList = this.userEntityService.findAll()
+                .stream()
+                .filter(user -> user.getStatus() == Status.DEACTIVE)
+                .toList();
+
+        return UserMapper.INSTANCE.convertToDTOs(userList);
+    }
+
+    @Override
+    public UserDTO findByIdInDeactive(Long id) {
+        User user = this.userEntityService.findByIdWithControl(id);
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
+        User userOptional = this.userEntityService.findAll()
+                .stream()
+                .filter(u -> u.getStatus() == Status.DEACTIVE && u.getId() == id)
+                .findFirst().orElseThrow();
+
+
+        return UserMapper.INSTANCE.convertToDTO(userOptional);
+    }
+
+    @Override
     public String delete(Long id) {
-        User user= this.userEntityService.findByIdWithControl(id);
+        User user = this.userEntityService.findByIdWithControl(id);
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
         this.userEntityService.delete(user);
         return Messages.USER_DELETED.getMessage();
     }
 
     @Override
     public UserDTO active(Long id) {
-        User user=this.userEntityService.findByIdWithControl(id);
+        User user = this.userEntityService.findByIdWithControl(id);
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
         this.userEntityService.changeStatusToActive(user.getId());
         return UserMapper.INSTANCE.convertToDTO(user);
     }
 
     @Override
     public UserDTO deactive(Long id) {
-        User user=this.userEntityService.findByIdWithControl(id);
+        User user = this.userEntityService.findByIdWithControl(id);
+        if (user == null) {
+            throw new ItemNotFoundException(ErrorMessages.NOT_FOUND_USER);
+        }
         this.userEntityService.changeStatusToDeactive(user.getId());
         return UserMapper.INSTANCE.convertToDTO(user);
     }
